+++ public/index.js
@@ -7,7 +7,12 @@ import '../src/plugin';
 
 window.onload = () => {
 	const video = videojs('my-video');
-	const zoomPlugin = video.zoomPlugin();
+	const zoomPlugin = video.zoomPlugin({
+		showZoom: true,
+		showMove: false,
+		showRotate: false,
+		gestureHandler: true
+	});
 	zoomPlugin.listen('click', () => {
 		console.log('onclick');
 	});